git-svn-id: https://groke.otago.ac.nz/ant/ant@6 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/memory.c
@@ -42,7 +42,7 @@ void *ans;
 	ans = at;
 	at += bytes;
 	if (at > chunk_end)
-		exit(printf("Out of memory:%d bytes requested %d bytes used", bytes, used));
+		exit(printf("Out of memory:%I64d bytes requested %I64d bytes used", (long long)bytes, used));
 	used += bytes;
 	}
 return ans;
